requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.0.1

# ─── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.28.1

# ─── Database ─────────────────────────────────────────────────────────────────
supabase==2.28.3
sqlalchemy==2.0.36
alembic==1.14.1
asyncpg==0.30.0
psycopg[binary]==3.2.4

# ─── Trend Analysis ───────────────────────────────────────────────────────────
pytrends==4.9.2
praw==7.8.1
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.49.1
feedparser==6.0.11

# ─── AI / LLM ─────────────────────────────────────────────────────────────────
groq

# ─── Voice Generation ─────────────────────────────────────────────────────────
edge-tts>=6.1.0
pydub>=0.25.1

# ─── Media / Video ────────────────────────────────────────────────────────────
moviepy==1.0.3
pillow==11.1.0
imageio==2.36.1
imageio-ffmpeg==0.5.1

# ─── Clip Extraction Pipeline ─────────────────────────────────────────────
yt-dlp>=2024.1.0
faster-whisper>=1.0.0
scenedetect[opencv]>=0.6.0
pyloudnorm>=0.1.1
opencv-python>=4.8.0

# ─── Subtitle Generation ──────────────────────────────────────────────────────
assemblyai>=0.35.0

# ─── Publishing APIs (YouTube only) ──────────────────────────────────────────
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1

# ─── File / Cloud Storage ─────────────────────────────────────────────────────
boto3==1.35.86
cloudflare==2.19.4

# ─── Scheduling / Orchestration ───────────────────────────────────────────────
celery>=5.3.0
apscheduler==3.10.4
schedule==1.2.2

# ─── Utilities ────────────────────────────────────────────────────────────────
python-slugify==8.0.4
arrow==1.3.0
tenacity==9.0.0
structlog==24.4.0
rich==13.9.4
typer==0.15.1
jinja2==3.1.5
aiofiles==24.1.0
python-multipart==0.0.20
pyyaml>=6.0.1

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
respx==0.22.0